improve story and folder structure
