+++ src/examples/SimpleWizard/SimpleWizard.stories.mdx
@@ -1,5 +1,6 @@
 import { Canvas, Meta, Story } from "@storybook/addon-docs";
-import * as FormioLoader from "../../helpers/FormioLoader";
+import { SimpleWizard } from "./SimpleWizard";
+import { simpleWizardCode } from "./SimpleWizard.code";
 
 <Meta title="Examples/SimpleWizard" />
 
@@ -7,45 +8,7 @@ import * as FormioLoader from "../../helpers/FormioLoader";
 
 This example show the default presentation of a wizard form.
 
-The demo can be reproduced with `FormioScript` to lazy load the Formio library and scripts.
-
-export function createSimpleWizard() {
-  const div = document.createElement("div");
-  setTimeout(() => {
-    FormioLoader.initFormio();
-  });
-  div.innerHTML = `
-        <div id="formio" 
-          class="qg-forms-v2"
-          data-formio 
-          data-formio-project-name="dev-svcwlpuksmwawwk" 
-          data-formio-form-name="simpleWizard" 
-          data-formio-env-url="api.forms.platforms.qld.gov.au" 
-        ></div>
-      `;
-  return div;
-}
-
-export const createSimpleWizardCode = `
-<head>
-  <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
-  <script src="https://static.qgov.net.au/formio-qld/v1/v1.x.x-latest/formio-script.prod.min.js"></script>
-</head>
-<body>
-  <script>
-${createSimpleWizard
-  .toString()
-  .replaceAll(
-    "_helpers_FormioLoader__WEBPACK_IMPORTED_MODULE_3__",
-    "FormLoader"
-  )}
-    FormioScript.init().then(() => {
-      const div = createSimpleWizard();
-      document.body.append(div);
-    });
-  </script>
-</body>
-`;
+The demo can be reproduced with [FormioScript](?path=/docs/helpers-formioscript--page) to lazy load the Formio library and scripts.
 
 <Canvas withSource="open">
   <Story
@@ -53,11 +16,11 @@ ${createSimpleWizard
     parameters={{
       docs: {
         source: {
-          code: createSimpleWizardCode,
+          code: simpleWizardCode,
         },
       },
     }}
   >
-    {() => createSimpleWizard()}
+    {() => SimpleWizard()}
   </Story>
 </Canvas>
