+++ src/examples/SingleSignOn/SingleSignOn.stories.mdx
@@ -1,9 +1,8 @@
 import { Canvas, Meta, Story } from "@storybook/addon-docs";
 import SSOHighLevel from "../../stories/assets/SSO-HighLevel.png";
 import SSOUserFlow from "../../stories/assets/SSO-UserFlow.png";
-import { createSSoForm } from "./SingleSignOn";
-// eslint-disable-next-line import/no-unresolved, import/extensions
-import { createSSoForm as createSSoFormRaw } from "!!./SingleSignOn";
+import { SingleSignOn } from "./SingleSignOn";
+import { singleSignOnCode } from "./SingleSignOn.code";
 
 <Meta title="Examples/SingleSignOn" />
 
@@ -35,28 +34,7 @@ Main Form : https://www.qld.gov.au/dev/steve/idb-poc-dev-use-only-queensland-bor
 
 The following demo shows the SSO form in a single page application.
 
-The demo can be reproduced with `FormioScript` to lazy load the Formio library and scripts.
-
-export const createSSoFormCode = `
-<head>
-  <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
-  <script src="https://static.qgov.net.au/formio-qld/v1/v1.x.x-latest/formio-script.prod.min.js"></script>
-</head>
-<body>
-  <script>
-${createSSoFormRaw
-  .toString()
-  .replaceAll(
-    "_helpers_FormioLoader__WEBPACK_IMPORTED_MODULE_0__",
-    "FormLoader"
-  )}
-    FormioScript.init().then(() => {
-      const div = createSSoForm();
-      document.body.append(div);
-    });
-  </script>
-</body>
-`;
+The demo can be reproduced with [FormioScript](?path=/docs/helpers-formioscript--page) to lazy load the Formio library and scripts.
 
 <Canvas withSource="open">
   <Story
@@ -64,12 +42,12 @@ ${createSSoFormRaw
     parameters={{
       docs: {
         source: {
-          code: createSSoFormCode,
+          code: singleSignOnCode,
           language: "html",
         },
       },
     }}
   >
-    {() => createSSoForm()}
+    {() => SingleSignOn()}
   </Story>
 </Canvas>
