+++ src/helpers/FormioLoader/stories/CustomController/CustomController.template.js
@@ -1,0 +1,3 @@
+export const customControllerTemplate = `
+<div id="formio-controller" class="qg-forms-v2" ></div>
+`;
