+++ src/helpers/FormioLoader/stories/CustomController/index.js
@@ -1,0 +1,3 @@
+export * from "./CustomController.code";
+export * from "./CustomController";
+export * from "./CustomController.template";
