+++ src/helpers/FormioLoader/stories/CustomOptions/CustomOptions.template.js
@@ -1,0 +1,3 @@
+export const customOptionsTemplate = `
+<div id="formio-options" class="qg-forms-v2" ></div>
+`;
