+++ src/helpers/FormioLoader/stories/CustomOptions/index.js
@@ -1,0 +1,3 @@
+export * from "./CustomOptions.code";
+export * from "./CustomOptions";
+export * from "./CustomOptions.template";
