+++ src/helpers/FormioLoader/stories/InitFormio/InitFormio.js
@@ -1,0 +1,5 @@
+import * as FormioLoader from "../..";
+
+export function InitFormio() {
+  FormioLoader.initFormio();
+}
