+++ src/helpers/FormioLoader/stories/InitFormio/InitFormio.template.js
@@ -1,0 +1,13 @@
+export const initFormioTemplate = `
+<div id="formio" 
+  class="qg-forms-v2"
+  data-formio 
+  data-formio-project-name="dev-svcwlpuksmwawwk" 
+  data-formio-form-name="plsPlusFormDemo" 
+  data-formio-form-confirmation="" 
+  data-formio-form-revision="" 
+  data-formio-env-url="api.forms.platforms.qld.gov.au" 
+  data-formio-pdf-download="no" 
+  data-formio-namespace="" 
+></div>
+`;
