+++ src/helpers/FormioLoader/stories/InitFormioInstance/InitFormioInstance.template.js
@@ -1,0 +1,3 @@
+export const initFormioInstanceTemplate = `
+<div id="formio-alt" class="qg-forms-v2" ></div>
+`;
