+++ src/helpers/FormioLoader/stories/InitFormioInstance/index.js
@@ -1,0 +1,3 @@
+export * from "./InitFormioInstance.code";
+export * from "./InitFormioInstance";
+export * from "./InitFormioInstance.template";
