+++ src/helpers/FormioScript/FormioScript.stories.mdx
@@ -7,6 +7,13 @@ import { Meta } from "@storybook/addon-docs";
 FormioScript provide shorthand implementation to lazy load the necessary libraries and scripts that need to initiate a Formio form in your application.
 Simply include it in your application and it will search through the page with Formio div container and initiated the Formio instances.
 
+**Benefits**
+
+- One script and for all, reduce the chance of missing dependency.
+- Enforce the script are loading in certain order, wrong order will break the application, which is prone to error if insert manually.
+- Always point to the correct source of scripts.
+- Encapsulated, avoid consumer input in case we need to change the composition of scripts.
+
 **Before using FormioScript:**
 
 ```html
